Convert from Blackboard XML to an R2BB test

This function converts an object from Blackboard XML format to an R2BB test.

Usage

from_bbxml_test(x, convert_html_to_markdown = TRUE, pools_by_name)

Arguments

x

An xml2 node or nodeset

convert_html_to_markdown

Whether to convert HTML to Markdown.

pools_by_name

A named list of pools, by resource name; used to fill in the pool details for random blocks.

Value

An R2BB test

Details

Generally you will not call this directly; instead, call from_bbxml_package() on an exported zip package file.

By default, rich text fields in HTML are converted to Markdown using pandoc. This can work quite well, but is not perfect; results are not guaranteed to be identical to the original.

See also