Convert a question pool to Blackboard XML

This function converts a question pool to Blackboard XML format. This is mostly used as part of to_bbxml_package(), but can be called directly if you want the XML for a single question.

Usage

# S3 method for class 'r2bb_pool'
to_bbxml(
  x,
  convert_rich_text = "md_to_html_pandoc",
  convert_rich_text_options = list(options = "--mathml"),
  ...
)

Arguments

x

The question pool to convert

convert_rich_text

Method to convert rich text to HTML. By default, md_to_html_pandoc is used, which converts Markdown to HTML using pandoc. Can be FALSE to prevent any conversion.

convert_rich_text_options

Additional arguments to pass to the convert_rich_text function.

...

Additional arguments passed to methods

Value

A character string containing the Blackboard XML

See also